+++ src/components/Login.jsx
@@ -1,0 +1,73 @@
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+
+function Login() {
+  const [username, setUsername] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const navigate = useNavigate();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError("");
+
+    try {
+      const response = await fetch("http://localhost:5002/api/login", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ username, email, password }),
+      });
+
+      if (response.ok) {
+        navigate("/todolist");
+      } else {
+        const errorMessage = await response.text();
+        setError(errorMessage);
+      }
+    } catch (err) {
+      console.error("エラー", err);
+      setError("ログイン中にエラーが発生しました");
+    }
+  };
+
+  return (
+    <div className="login-container">
+      <h2>ログイン</h2>
+      {error && <div className="error">{error}</div>}
+      <form onSubmit={handleSubmit}>
+        <input
+          type="text"
+          placeholder="ユーザー名"
+          value={username}
+          onChange={(e) => setUsername(e.target.value)}
+          required
+        />
+        <input
+          type="email"
+          placeholder="メールアドレス"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+        />
+        <input
+          type="password"
+          placeholder="パスワード"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+        />
+        <div className="button-container">
+          <button type="submit">ログイン</button>
+        </div>
+      </form>
+      <div className="button-container">
+        <button onClick={() => navigate("/signup")}>サインアップ画面へ</button>
+      </div>
+    </div>
+  );
+}
+
+export default Login;
