+++ src/components/TodoList.jsx
@@ -1,95 +1,122 @@
-import { useState } from "react";
-import PropTypes from "prop-types";
+import { useState, useEffect } from "react";
+import TodoForm from "./TodoForm";
+import TodoTask from "./TodoTask";
 
-function TodoList({ todos, toggleComplete, deleteTodo, updateTodoText }) {
+function TodoList() {
+  const [todos, setTodos] = useState([]);
+  const [error, setError] = useState(null);
+  const BASE_URL = "http://localhost:5002";
+
+  useEffect(() => {
+    const fetchTodos = async () => {
+      try {
+        const response = await fetch(`${BASE_URL}/api/todos`);
+        if (!response.ok) {
+          throw new Error("Failed to fetch todos");
+        }
+        const data = await response.json();
+        setTodos(data);
+      } catch (error) {
+        console.error("Error fetching todos:", error);
+        setError("タスクの取得に失敗しました。");
+      }
+    };
+    fetchTodos();
+  }, []);
+
+  async function addTodo(text) {
+    const newTodo = { text, completed: false };
+    try {
+      const response = await fetch(`${BASE_URL}/api/todos`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(newTodo),
+      });
+      if (!response.ok) {
+        throw new Error("Failed to add todo");
+      }
+      const createdTodo = await response.json();
+      setTodos([...todos, createdTodo]);
+    } catch (error) {
+      console.error("Error adding todo:", error);
+      setError("タスクの追加に失敗しました。");
+    }
+  }
+
+  async function toggleComplete(id) {
+    const updatedTodos = todos.map((todo) =>
+      todo.id === id ? { ...todo, completed: !todo.completed } : todo
+    );
+    setTodos(updatedTodos);
+    try {
+      const response = await fetch(`${BASE_URL}/api/todos/${id}`, {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          completed: updatedTodos.find((todo) => todo.id === id).completed,
+        }),
+      });
+      if (!response.ok) {
+        throw new Error("Failed to update todo");
+      }
+    } catch (error) {
+      console.error("Error updating todo:", error);
+      setError("タスクの更新に失敗しました。");
+    }
+  }
+
+  async function updateTodoText(id, newText) {
+    try {
+      const response = await fetch(`${BASE_URL}/api/todos/${id}`, {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ text: newText }),
+      });
+      if (!response.ok) {
+        throw new Error("Failed to update todo text");
+      }
+      const updatedTodo = await response.json();
+      setTodos(todos.map((todo) => (todo.id === id ? updatedTodo : todo)));
+    } catch (error) {
+      console.error("Error updating todo text:", error);
+      setError("タスクの内容更新に失敗しました。");
+    }
+  }
+
+  async function deleteTodo(id) {
+    try {
+      const response = await fetch(`${BASE_URL}/api/todos/${id}`, {
+        method: "DELETE",
+      });
+      if (!response.ok) {
+        throw new Error("Failed to delete todo");
+      }
+      setTodos(todos.filter((todo) => todo.id !== id));
+    } catch (error) {
+      console.error("Error deleting todo:", error);
+      setError("タスクの削除に失敗しました。");
+    }
+  }
+
   return (
-    <ul>
-      {todos.map((todo, index) => (
-        <TodoItem
-          key={`${todo.id}_${index}`}
-          todo={todo}
-          toggleComplete={toggleComplete}
-          deleteTodo={deleteTodo}
-          updateTodoText={updateTodoText}
-        />
-      ))}
-    </ul>
+    <div className="app">
+      <h1>TODOリスト</h1>
+      {error && <p className="error">{error}</p>}
+      <TodoForm addTodo={addTodo} />
+      <TodoTask
+        todos={todos}
+        toggleComplete={toggleComplete}
+        deleteTodo={deleteTodo}
+        updateTodoText={updateTodoText}
+      />
+    </div>
   );
 }
-
-function TodoItem({ todo, toggleComplete, deleteTodo, updateTodoText }) {
-  const [isEditing, setIsEditing] = useState(false);
-  const [newText, setNewText] = useState(todo.text);
-
-  const handleEditToggle = () => {
-    setIsEditing(!isEditing);
-    setNewText(todo.text);
-  };
-
-  const handleUpdate = async () => {
-    await updateTodoText(todo.id, newText);
-    setIsEditing(false);
-  };
-
-  return (
-    <li>
-      {isEditing ? (
-        <input
-          type="text"
-          value={newText}
-          onChange={(e) => setNewText(e.target.value)}
-          onBlur={handleUpdate}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              handleUpdate();
-            }
-          }}
-        />
-      ) : (
-        <span
-          style={{
-            textDecoration: todo.completed ? "line-through" : "none",
-          }}
-          onClick={handleEditToggle}
-        >
-          {todo.text}
-        </span>
-      )}
-      <div className="task-buttons">
-        <input
-          type="checkbox"
-          checked={todo.completed}
-          onChange={() => toggleComplete(todo.id)}
-          className="checkbox"
-        />
-        <button onClick={() => deleteTodo(todo.id)}>削除</button>
-      </div>
-    </li>
-  );
-}
-
-TodoList.propTypes = {
-  todos: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      text: PropTypes.string.isRequired,
-      completed: PropTypes.bool.isRequired,
-    })
-  ).isRequired,
-  toggleComplete: PropTypes.func.isRequired,
-  deleteTodo: PropTypes.func.isRequired,
-  updateTodoText: PropTypes.func.isRequired,
-};
-
-TodoItem.propTypes = {
-  todo: PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    text: PropTypes.string.isRequired,
-    completed: PropTypes.bool.isRequired,
-  }).isRequired,
-  toggleComplete: PropTypes.func.isRequired,
-  deleteTodo: PropTypes.func.isRequired,
-  updateTodoText: PropTypes.func.isRequired,
-};
 
 export default TodoList;
