+++ src/index.css
@@ -32,15 +32,51 @@ li {
   font-weight: bold;
 }
 
+.login-container,
+.signup-container {
+  width: 300px;
+  padding: 20px;
+  background-color: #ffffff;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+h2 {
+  text-align: center;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+input[type="text"],
+input[type="email"],
+input[type="password"] {
+  padding: 8px;
+  border-radius: 4px;
+  border: 1px solid #ccc;
+}
+
+button {
+  flex: 1;
+  padding: 10px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  margin-top: 5px;
+}
+
 .task-buttons {
   display: flex;
   gap: 0.5rem;
 }
 
-button,
 .checkbox {
-  width: 80px;
-  height: 30px;
   border: 1px solid black;
   border-radius: 4px;
   background-color: #007bff;
@@ -49,32 +85,22 @@ button,
   transition: background-color 0.3s;
 }
 
+.button-container {
+  display: flex;
+  gap: 10px;
+}
+
 button:hover,
 .checkbox:hover {
   background-color: #0056b3;
 }
 
-input[type="text"] {
-  border: 1px solid black;
-  border-radius: 4px;
-  padding: 0.25rem 0.5rem;
-  font-size: 1rem;
-  flex-grow: 1;
-  margin-right: 10px;
-}
-
-input[type="checkbox"] {
-  width: 20px;
-  height: 20px;
-  margin-right: 10px;
-}
-
-.completed {
-  text-decoration: line-through;
-  color: #888;
-}
-
 .error {
   color: red;
+  margin-bottom: 1rem;
+}
+
+.message {
+  color: green;
   margin-bottom: 1rem;
 }
